Sly State Machine + Jump Fixes
Refactored Sly's code into a defined state machine and wrote some notes for future changes.

Added acceleration to movement and played with it to make the jumps feel closer to OG Sly games.

Gave the ledge grab an automatic version when Sly is falling downward and is close to the ledge.
You can still press circle to grab ledges for now, but that's subject to change later on.