Made Sly's Tween Smoother
In move_to_target, I created a tween for each axis:
x_tween
y_tween
z_tween

Set the y tween to CUBIC and the x and z to LINEAR to create a curve in his movement toward platforms. This only works on ledges and spire jumps at the moment, but I'll add it to paths later on.