Jump Corrections
Sly's mesh moves to the bottom of his collision shape as he jumps. Some of the temporary jump animations place Sly's mesh too high, leaving a gap between the collision shape and his mesh. Moving his mesh down as he jumps closes this gap and gives the player a better understanding of where they're at in the air.

If we get custom animations, this can be accounted for in those instead of the code.