Enemy AI Follow and Detection
Added a Spotlight scene that detects the player if they're in range and aren't behind an object (just an area node, spotlight, and raycast). This scene will control the AI's direction and rotation down the line, so they will always follow the spotlight.

The new Enemy AI scene will handle states and act as the parent of smaller nodes (spotlight, gun, etc.) that the enemy uses. It also holds an exported variable called Nav Parent that assigns five points for the enemy to randomly move to during its IDLE state. When the enemy reaches its point, a new one will be assigned, and its path will change.

I also edited the jump and air animations. Sly's feet are now at the bottom of his collision shape during these animations, and they're much smoother.

Everything is WIP