Enemy Communication (Rough)
Enemies will communicate with each other to detect the player.

If an enemy detects another enemy that is chasing, they will go to that enemy (yellow sphere above their head). When they get close enough to them, they will target the player and switch to chase mode.

Need to add: When Enemy spots the player with their flashlight and another enemy is in range, they should be alerted. (Note: This works when the player is detected by sprinting, just not with a flashlight)

A good start, I think