Wall Notch Magnetism 1.0
Added the Wall Notch thief move.

When Sly is in the air and in range of a wall notch, he will be slightly magnetized to it. His jumps from wall notch points are higher than normal jumps.

Wall notches shouldn't be within range of the floor to prevent unwanted magnetism (can't be fully fixed by only applying magnetism in the air) so be careful of where you place the node.

Strength, range, and jump height are all WIP. I'm also looking for ways to optimize this.