Improvements: Camera, Tail, Paraglider
Camera has new conditions that make it more reliable, especially notable when Sly is falling.

Tail sways more and is more consistent. There's an issue when Sly is doing a ledge grab and a minor issue when he's on a pipe, but this is much better than it was previously.

Paraglider reacts more quickly and is closer to the original games.